workflows/build: test Portable Ruby in home directory
For Linux builds with load-relative enabled, Ruby gets its runtime
location through `readlink` on `/proc/self/exe` [^1]. Internally,
`readlink` is called repeatedly with exponentially increasing buffer
sizes until the path is successfully read [^2].

On aarch64, `readlink` is implemented indirectly via the `readlinkat`
system call, because the `readlink` system call is absent [^3]. Due to a
bug in QEMU, however, `readlinkat` incorrectly returns the full length
of the result regardless of the buffer size [^4]. If QEMU is run with
QEMU_STRACE set, the following can be seen:

    readlinkat(AT_FDCWD,"/proc/self/exe",0x0000005500da8a00,100) = 114

The buffer size is 100, so should the result length be longer, the
system call ought to return the number of bytes actually written to the
buffer (i.e., 100). In the example above, however, a greater number is
incorrectly returned. This causes Ruby to incorrectly determine that
the buffer is large enough to hold the path, and thus fail to correctly
determine its runtime location and load the required gems.

This has been fixed in QEMU 7.2.0 [^5] [^6]. However, since this seems
to be the only QEMU issue that affects the test as of now, to avoid
adding maintainance burden regarding QEMU updates, we can choose to work
around it by testing the build at the home directory instead, where the
path is shorter. This is not a problem with the build.

[^1]: https://github.com/ruby/ruby/blob/v3_1_4/ruby.c#L635
[^2]: https://github.com/ruby/ruby/blob/v3_1_4/file.c#L3153-L3180
[^3]: https://sourceware.org/git/?p=glibc.git;a=blob;f=sysdeps/unix/sysv/linux/readlink.c;h=bb75d204170f95f90ad15e60590e0e46fe73286c;hb=41903cb6f460d62ba6dd2f4883116e2a624ee6f8
[^4]: https://lists.nongnu.org/archive/html/qemu-devel/2022-08/msg01389.html
[^5]: https://lists.nongnu.org/archive/html/qemu-devel/2022-09/msg03674.html
[^6]: https://github.com/qemu/qemu/commit/65d4830dac7d88059ba77f1d31ec0c2f8f65ae28